wokring on reservation weblink

+++ app/Controllers/ApplicatioController.php
@@ -2981,7 +2981,7 @@ class ApplicatioController extends BaseController
         }
     }
     public function webLineReservation($resvid){
-        try{
+        // try{
             $param = ['RESV_ID'=> $resvid];
             $sql = "SELECT RESV_ID, RESV_NO, FORMAT(RESV_ARRIVAL_DT,'dd-MMM-yyyy') RESV_ARRIVAL_DT, 
                         RESV_NIGHT, RESV_ADULTS, RESV_CHILDREN, FORMAT(RESV_DEPARTURE,'dd-MMM-yyyy') RESV_DEPARTURE,
@@ -2992,11 +2992,15 @@ class ApplicatioController extends BaseController
                         CUST_EMAIL, CUST_MOBILE, CUST_PHONE, CUST_POSTAL_CODE, CUST_NATIONALITY, CUST_DOC_TYPE, 
                         CUST_GENDER, CUST_DOC_EXPIRY, CUST_DOC_NUMBER, FORMAT(CUST_DOC_ISSUE,'dd-MMM-yyyy') CUST_DOC_ISSUE,
                         RESV_ACCP_TRM_CONDI, RESV_SINGATURE_URL, RESV_ETA,
-                        (SELECT RM_TY_DESC FROM FLXY_ROOM_TYPE WHERE RM_TY_CODE=RESV_RM_TYPE) RM_TY_DESC, 
-                        (SELECT SNAME FROM STATE WHERE STATE_CODE = CUST_STATE AND COUNTRY_CODE = CUST_COUNTRY) CUST_STATE_DESC, 
-                        (SELECT CTNAME FROM CITY WHERE ID = CUST_CITY) CUST_CITY_DESC
-                        FROM FLXY_RESERVATION, FLXY_CUSTOMER 
-                        WHERE RESV_ID = :RESV_ID: AND RESV_NAME = CUST_ID";
+                        RM_TY_DESC,
+                        SNAME as CUST_STATE_DESC,
+                        CTNAME as CUST_CITY_DESC 
+                        FROM FLXY_RESERVATION
+                        left join FLXY_CUSTOMER on FLXY_RESERVATION.RESV_NAME = FLXY_CUSTOMER.CUST_ID
+                        left join FLXY_ROOM_TYPE on FLXY_RESERVATION.RESV_RM_TYPE = FLXY_ROOM_TYPE.RM_TY_CODE
+                        left join STATE on STATE.STATE_CODE = FLXY_CUSTOMER.CUST_STATE
+                        left join CITY on CITY.ID = FLXY_CUSTOMER.CUST_CITY
+                        WHERE RESV_ID = :RESV_ID:";
 
             $response = $this->Db->query($sql,$param)->getResultArray();
 
@@ -3009,19 +3013,19 @@ class ApplicatioController extends BaseController
             $data['condition'] = '';
 
             return view('WebCheckin/CheckInReservation', $data);
-        }catch (\Exception $e){
-            return $this->respond($e->errors());
-        }
+        // }catch (\Exception $e){
+            // return $this->respond($e->getMessage());
+        // }
     }
 
     function imageUpload(){
         try{
             $avatar = $this->request->getFile('file');
-            $avatar->move(ROOTPATH . 'assets/Uploads/');
+            $avatar->move(ROOTPATH . 'assets/Uploads/UserDocuments/proof/');
             echo $avatar->getClientName();
             exit;
-        }catch (Exception $e){
-            return $this->respond($e->errors());
+        }catch (\Exception $e){
+            return $this->respond($e->getMessage());
         }
     }
 
@@ -3038,13 +3042,13 @@ class ApplicatioController extends BaseController
                 $dst_r = ImageCreateTrueColor( $_POST['w'], $_POST['h'] );
                 imagecopyresampled($dst_r, $img_r, 0, 0, $_POST['x'], $_POST['y'], $_POST['w'], $_POST['h'], $_POST['w'],$_POST['h']);
                 // header('Content-type: image/jpeg');
-                imagejpeg($dst_r,'assets/Uploads/userDocuments/'.$newFile);
+                imagejpeg($dst_r,'assets/Uploads/UserDocuments/proof/'.$newFile);
             }else{
-                $sourcePath = dirname(__DIR__,2). '/assets/Uploads/userDocuments/'.$imageName;
-                $newPath = dirname(__DIR__,2). '/assets/Uploads/userDocuments/'.$newFile;
+                $sourcePath = dirname(__DIR__,2). '/assets/Uploads/UserDocuments/proof/'.$imageName;
+                $newPath = dirname(__DIR__,2). '/assets/Uploads/UserDocuments/proof/'.$newFile;
                 rename($sourcePath,$newPath);
             }
-            $file_unlink = dirname(__DIR__,2).'/assets/Uploads/userDocuments/'.$imageName;
+            $file_unlink = dirname(__DIR__,2).'/assets/Uploads/UserDocuments/proof/'.$imageName;
             if(file_exists($file_unlink)){
                 unlink($file_unlink);
             }
@@ -3057,16 +3061,29 @@ class ApplicatioController extends BaseController
                 "DOC_CREATE_UID" => $this->session->name,
                 "DOC_CREATE_DT" => $dateTime
             ];
-            $return = $this->Db->table('FLXY_DOCUMENTS')->insert($data);
+
+            $uploaded_documents = $this->Db->table('FLXY_DOCUMENTS')->where('DOC_CUST_ID', $this->request->getPost("DOC_CUST_ID"))
+                ->where('DOC_RESV_ID', $this->request->getPost("DOC_RESV_ID"))->get()->getResultArray();
+            
+            if(count($uploaded_documents)){
+                $return = $this->Db->table('FLXY_DOCUMENTS')->where('DOC_CUST_ID', $this->request->getPost("DOC_CUST_ID"))
+                                    ->where('DOC_RESV_ID', $this->request->getPost("DOC_RESV_ID"))
+                                    ->update(['DOC_FILE_PATH' => $uploaded_documents[0]['DOC_FILE_PATH'] . ',' . $newFile]);
+            }
+            else{
+                $return = $this->Db->table('FLXY_DOCUMENTS')->insert($data);
+            }
+
             if($return){
                 $array = array('DOC_CUST_ID' => $this->request->getPost("DOC_CUST_ID"), 'DOC_RESV_ID' => $this->request->getPost("DOC_RESV_ID"));
                 $listImage = $this->Db->table('FLXY_DOCUMENTS')->select('DOC_ID,DOC_FILE_PATH,DOC_FILE_TYPE')->where($array)->orderBy('DOC_ID', 'DESC')->get()->getResultArray();
-                $imagePath = array('IMAGEPATH'=>$newFile);
-                $outPut = array_merge($imagePath,$listImage);
+                // $imagePath = array('IMAGEPATH'=>$newFile);
+                // $outPut = array_merge($imagePath,$listImage);
+                $outPut = $listImage[0];
                 $result = $this->responseJson("1","0",$return,$outPut);
                 echo json_encode($result);
             }else{
-                $result = $this->responseJson("-444",$message,$return);
+                $result = $this->responseJson("-444", $message, $return);
                 echo json_encode($result);
             }
         }catch (Exception $e){
@@ -3082,13 +3099,29 @@ class ApplicatioController extends BaseController
 
     public function deleteUploadImages(){
         try{
+            $return = false;
+
             $sysid = $this->request->getPost("sysid");
-            $return = $this->Db->table('FLXY_DOCUMENTS')->delete(['DOC_ID' => $sysid]); 
+            $file_name = $this->request->getPost("file_name");
+
+            $uploaded_documents = $this->Db->table('FLXY_DOCUMENTS')->where('DOC_ID', $sysid)->get()->getResultArray(); 
+            if(count($uploaded_documents)){
+                $uploaded_documents = $uploaded_documents[0];
+                $uploaded_documents_arr = explode(",", $uploaded_documents['DOC_FILE_PATH']);
+                
+                if(count($uploaded_documents_arr) == 1){
+                    $return = $this->Db->table('FLXY_DOCUMENTS')->delete(['DOC_ID' => $sysid]); 
+                }
+
+                $return = $this->Db->table('FLXY_DOCUMENTS')->where('DOC_ID', $sysid)
+                                    ->update(['DOC_FILE_PATH' => str_replace(",$file_name", '', $uploaded_documents['DOC_FILE_PATH'])]); 
+            }
+
             if($return){
                 $result = $this->responseJson("1","0",$return,$response='');
                 echo json_encode($result);
             }else{
-                $result = $this->responseJson("-444",$message,$return);
+                $result = $this->responseJson("-444", 'No documents', $return);
                 echo json_encode($result);
             }
         }catch (Exception $e){
@@ -3098,7 +3131,9 @@ class ApplicatioController extends BaseController
 
     public function updateCustomerDetail(){
         try{
+            $resvid = $this->request->getPost("DOC_RESV_ID");
             $custId = $this->request->getPost("DOC_CUST_ID");
+
             $data = ["CUST_TITLE" => $this->request->getPost("CUST_TITLE"),
                 "CUST_FIRST_NAME" => $this->request->getPost("CUST_FIRST_NAME"),
                 "CUST_LAST_NAME" => $this->request->getPost("CUST_LAST_NAME"),
@@ -3119,16 +3154,17 @@ class ApplicatioController extends BaseController
                 "CUST_UPDATE_DT" => date("Y-m-d")
             ];
             $return = $this->Db->table('FLXY_CUSTOMER')->where('CUST_ID', $custId)->update($data); 
+
             if($return){
                 $response = $this->checkStatusUploadFiles('RTN',$custId,$resvid);
                 $result = $this->responseJson("1","0",$return,$response);
                 echo json_encode($result);
             }else{
-                $result = $this->responseJson("-444",$message,$return);
+                $result = $this->responseJson("-444", 'unable to update customer data.', $return);
                 echo json_encode($result);
             }
-        }catch (Exception $e){
-            return $this->respond($e->errors());
+        }catch (\Exception $e){
+            return $this->respond($e->getMessage());
         }
     }
 
@@ -3156,7 +3192,7 @@ class ApplicatioController extends BaseController
                     $fileArry=$this->request->getFileMultiple('files');
                     foreach($fileArry as $key=>$file){ 
                         $newName = $file->getRandomName();
-                        $file->move(ROOTPATH . 'assets/Uploads/userDocuments/vaccination',$newName);
+                        $file->move(ROOTPATH . 'assets/Uploads/UserDocuments/vaccination',$newName);
                         $comma='';
                         if (isset($fileArry[$key+1])) {
                             $comma=',';
@@ -3232,7 +3268,7 @@ class ApplicatioController extends BaseController
                 $fileName = time();
                 $signature=base64_decode($parts[1]);
                 $fileNameExt = $fileName.'.'.$extension;
-                file_put_contents('assets/Uploads/userDocuments/signature/'.$fileNameExt,$signature);
+                file_put_contents('assets/Uploads/UserDocuments/signature/'.$fileNameExt,$signature);
             }else{
                 $fileNameExt = basename($signature);   
             }
